requirements.txt
requests==2.22.0
requests-html==0.10.0
python-magic==0.4.15